backend/requirements.txt
flask
flask-cors
sentence-transformers
faiss-cpu
numpy
Flask-CORS==4.0.0
Werkzeug==2.3.7
